==> Code/1_Script_Preprocessing_JoinDatasets.ipynb <==
## Import Libraries


import pandas as pd
import os
import json
import numpy as np

# connect to google drive
from google.colab import drive
drive.mount('/content/drive')

## Data Loading

Load all json files in one df (for further processing)

# Create an empty list to store data
data = []

# Specify the path where your JSON files are located
paths =  ['/content/drive/MyDrive/NLP/Data3/', '/content/drive/MyDrive/NLP/Data2/']


# Iterate over all paths
for path in paths:
    # Iterate over all files in the specified directory
    for filename in os.listdir(path):
        # Check if the file is a JSON file
        if filename.endswith('.json'):
            # Full file path
            file_path = os.path.join(path, filename)
            # Load JSON file
            with open(file_path, 'r', encoding='utf-8') as f:
                json_data = json.load(f)
            # Add data from the current file to the overall data list
            data.extend(json_data)

# Convert list into a DataFrame
df = pd.DataFrame(data)
# Create new columns for 'lat' and 'lng'
df['lat'] = df['location'].apply(lambda x: x.get('lat') if isinstance(x, dict) else np.nan)
df['lng'] = df['location'].apply(lambda x: x.get('lng') if isinstance(x, dict) else np.nan)

# Drop irrelevant columns
df = df[["reviews", 'url', 'name', 'stars', 'numberOfGuests', 'address',
       'pricing', 'primaryHost', 'isHostedBySuperhost', 'lat', 'lng']]

# For reviews: create seperate df
df_reviews = pd.json_normalize(data, 'reviews', ['url', 'name', 'stars', 'numberOfGuests', "isHostedBySuperhost"], record_prefix='review_', errors = "ignore")
# Drop irrelevant columns
df_reviews = df_reviews.drop(columns=["review_collectionTag", "review_response", "review_localizedReview", "review_author.hasProfilePic", "review_author.pictureUrl", 'review_recipient.hasProfilePic', "review_recipient.pictureUrl", 'review_author.firstName', 'review_author.smartName', 'review_author.thumbnailUrl','review_recipient.firstName', 'review_recipient.id','review_recipient.smartName', 'review_recipient.thumbnailUrl','review_localizedReview.comments','review_localizedReview.needsTranslation', "review_localizedReview.response"])

# Drop "review" from df (use df_reviews from now on - df only to have additional info about airbnb)
df = df.drop(columns=["reviews"])

## Save to OneDrive

# Save dataframes at google drive
df_reviews.to_csv('/content/drive/MyDrive/NLP/Handin/Data/df_reviews_uncleaned.csv', index=False)
df.to_csv('/content/drive/MyDrive/NLP/Handin/Data/df_uncleaned.csv', index=False)

## Overview combined Dataset

# to load dataframes from google drive
df = pd.read_csv('/content/drive/MyDrive/NLP/Handin/Data/df_uncleaned.csv')
df_reviews = pd.read_csv('/content/drive/MyDrive/NLP/Handin/Data/df_reviews_uncleaned.csv')

# Number of columns
print("Number of uncleaned Airbnbs: ", len(df))
print("Number of uncleaned Reviews: ", len(df_reviews))

# Look at columns of dfs
print(df.columns)
print(df_reviews.columns)

# Show fist 5 rows of dataframe storing Airbnb info
print(df.head())

# Show fist 5 rows of dataframe storing Review info
print(df_reviews.head())
