==> Code/3_Additional_Preprocessing.ipynb <==
# Pre-Processing for Models

### Import relevant libraries

# import relevant libraries

# for data manipulation
import numpy as np
import pandas as pd

# for NLP
import nltk
from nltk import pos_tag
from nltk import ngrams
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer
from nltk.stem import PorterStemmer
from nltk.corpus import stopwords
from collections import defaultdict
import re

# for counter
from tqdm import tqdm

### Import data before pre-processing

# import the data
df = pd.read_csv('df_reviews_final_binary.csv')

# check shape
df.shape

# are there reviews with no words
df['review_comments'].isna().any()

# check the ratings
df['review_rating'].value_counts()

# assign review column to a new dataframe
data = df['review_comments']

### Pre-processing of the data

##### Already included pre-processing steps

In the first pre-processing script, the removal of special characters and lowercasing was considered. Also using named entity recognition (NER), names and locations have been filtered as these are highly diverse and not contributing much to the sentiment. Also, this reduces the final resulting feature space of the document-term matrix and eases computation. 

Further pre-processing steps for the models is lemmatization, stemming, and stopwords removal to further decrease the feature space of the document-term matrix. We take unigrams, bigrams, and trigrams into account.

##### Tokenization (list of lists)

# if tokenizer not installed yet
import nltk
nltk.download('punkt')

# tokenize with list comprehension
tokenized_texts = [word_tokenize(text) for text in tqdm(data)]

tokenized_texts[0]

##### Lemmatization of words

# if lemmatizer not installed yet
import nltk
nltk.download('wordnet')

# if pos tagger not installed yet
import nltk
nltk.download('averaged_perceptron_tagger')

# initiate lemmatizer
lemmatizer = WordNetLemmatizer()

lemmatized_texts = []

# loop 
for texts in tqdm(tokenized_texts):
    tokens = []
    for token in texts:
        # pos_tag is needed for lemmatization
        tag = pos_tag([token])[0][1][0].lower()
        if tag == 'j':
            tag = 'a'
        # if word belongs to certain class, lemmatization is considered
        if tag in ('n','v','r','a'):
            tokens.append(lemmatizer.lemmatize(token, tag))
        # if not append word without processing
        else:
            tokens.append(token)
    lemmatized_texts.append(tokens)

lemmatized_texts[0]

##### Stemming of words

# initiate stemmer
stemmer = PorterStemmer()

# apply stemming on the words
stemmed_texts = [[stemmer.stem(token) for token in token_list] for token_list in tqdm(lemmatized_texts)]

##### Remove stopwords

# if stopwords not installed yet
import nltk
nltk.download('stopwords')

# save stopwords
stopwords = stopwords.words('english')

# look at some stopwords
print(stopwords[:5])

# some stopwords that can be interpreted negatively should be kept (negators)
negative_words = ['not','no','can','don','won','shouldn','couldn','wouldn','doesn','isn','aren','wasn','weren','but','nor']

# remove stopwords with list comprehension but keep the negative words
final_tokens = [[token for token in token_list if (token not in stopwords or token in negative_words)] for token_list in tqdm(stemmed_texts)]

final_tokens[0]

##### Remove rare words (to reduce dimension of document-term matrix later)

# count appearance of words and put it into dictionary

word_counts = defaultdict(int)

for tokens in final_tokens:
    for token in tokens:
        word_counts[token] += 1

# check amount of vocabulary
len(word_counts.keys())

# function that removes very rare words
def remove_infrequent_words(tokenized_documents, min_frequency=10):
    word_counts = defaultdict(int)

    # count word frequencies
    for tokens in tokenized_documents:
        for token in tokens:
            word_counts[token] += 1

    # filter infrequent words
    frequent_words = set()
    for word, count in word_counts.items():
        if count >= min_frequency:
            frequent_words.add(word)

    # remove rare words
    filtered_documents = []
    for tokens in tokenized_documents:
        filtered_tokens = [token for token in tokens if token in frequent_words]
        filtered_documents.append(filtered_tokens)

    return filtered_documents

# filter with the function
filtered = remove_infrequent_words(final_tokens, 10)

filtered[0]

# check how many unique words there are
word_counts = defaultdict(int)

for tokens in filtered:
    for token in tokens:
        word_counts[token] += 1

# check amount of vocabulary
len(word_counts.keys())

##### Extract Unigrams, Bigrams and Trigrams 

# function to extract ngrams from a list of tokens
def extract_ngrams(tokens, n=1):
    return ['_'.join(gram) for gram in ngrams(tokens, n)]

# extracting unigrams, bigrams, and trigrams
ngrams_list = []
for text in tqdm(filtered):
    text_ngrams = []
    text_ngrams += extract_ngrams(text, n=1)
    text_ngrams += extract_ngrams(text, n=2)
    text_ngrams += extract_ngrams(text, n=3)
    ngrams_list.append(text_ngrams)

ngrams_list[0]

##### Remove rare words after unigram, bigram, and trigram extraction again (to reduce dimension of document-term matrix later)

# count appearance of words and put it into dictionary

word_counts = defaultdict(int)

for tokens in ngrams_list:
    for token in tokens:
        word_counts[token] += 1

# check amount of vocabulary
len(word_counts.keys())

# filter with the function
final = remove_infrequent_words(ngrams_list, 10)

final[0]

# check how many unique words there are
word_counts = defaultdict(int)

for tokens in final:
    for token in tokens:
        word_counts[token] += 1

# check amount of vocabulary
len(word_counts.keys())

##### Create dataframe

# combine the final_tokens list and sentiment labels
df = pd.DataFrame(
    {'tokens': final,
     'sentiment': df['review_rating']
    })

# check dataframe
df.head()

# check that there are no empty tokens lists
df[df['tokens']=='[]']

### Save / load the data after processing  

# save df as csv
df.to_csv('df_tokenized.csv', index=False)

# load data as df
# df = pd.read_csv('df_tokenized.csv', index_col=[0])
