==> Code/4_Topic_Modeling.ipynb <==
# Topic Modeling

### Load relevant libraries and data

# for data manipulation
import numpy as np
import pandas as pd
import ast

# for NLP
import nltk
from nltk.corpus import stopwords
import gensim
import gensim.corpora as corpora
from gensim import models
from gensim.utils import simple_preprocess
from gensim.models import CoherenceModel
from gensim.models import TfidfModel
import spacy
import pyLDAvis
import pyLDAvis.gensim

# for counter
from tqdm import tqdm

# load the already pre-processed dataframe
df = pd.read_csv('df_tokenized.csv')

df.head()

In following steps only nouns, adjectives, verbs and adverbs are taken into account!

# convert to list of lists

# get back tokenized form
data = df['tokens'].apply(ast.literal_eval)

# convert back to a list of lists
data = data.tolist()

# initiate spacy pos tagging
nlp = spacy.load('en_core_web_sm')

# only including nouns, adjectives, verbs and adverbs
final_data = []

for tokens in tqdm(data):
    review = nlp(' '.join(tokens))
    new_rev = []
    for word in review:
        if word.pos_ in ['NOUN', 'ADJ', 'VERB', 'ADV']:
            new_rev.append(word)
    final_data.append(new_rev)

# convert spacy format back to string
data = [[str(token) for token in tokens] for tokens in tqdm(final_data)]

# convert to dataframe
df = pd.DataFrame(
    {'tokens': data,
     'sentiment': df['sentiment']
    })

# save df as csv
# df.to_csv('df_LDA.csv', index=False)

# load the data
df = pd.read_csv('topic_modeling_final.csv')

# convert to list of lists

# get back tokenized form
data = df['tokens'].apply(ast.literal_eval)

# convert back to a list of lists
data = data.tolist()

data[:2]

### Data pre-processing

Tokenization, lemmatization, stemming have already been done. The following code does the other steps necessary for topic modeling.

##### Create dictionary with all unique words: id2word

# create id2word which contains all unique words used in all reviews
id2word = corpora.Dictionary(data)

# create a list containing all reviews with their respective used words and word count
corpus = []
for text in data:
    new = id2word.doc2bow(text)
    corpus.append(new)

# check the first review
print(corpus[0])

However, we will filter for extreme values later to get a better result and create a new list that will replace this one.

# check that length of corpus is as long as the amount of documents
len(corpus)

# remove duplicate documents
corpus = [x for x in corpus if x]

# check new length of corpus
len(corpus)

##### Filter for extreme values

Tokens that appear in less than 30 reviews are filtered out or in more than 80% of reviews will be filtered out.

Finally after both first steps, we keep the 10,000 most frequent tokens.

We played with the thresholds / parameters to obtain a satisfying result.

# filter for extreme values
id2word.filter_extremes(no_below=30, no_above=0.8, keep_n=10000)

# check new corpus length
print(len(id2word.iteritems()))

##### Create a list containing all reviews with their respective used words and word count: bow_corpus

bow_corpus = [id2word.doc2bow(doc) for doc in data]

# check the first review
print(bow_corpus[0])

# check that length of corpus is as long as the amount of documents
print(len(bow_corpus))

# check the representation with actual words
first_review = bow_corpus[0]

for token in range(len(first_review)):
    print('Word {} (\"{}\") appears {} time.'.format(first_review[token][0], id2word[first_review[token][0]], first_review[token][1]))

### Apply tf-idf scores to the corpus

# apply tf-idf to the bow_corpus
tfidf_model = TfidfModel(bow_corpus)
corpus_tfidf = tfidf_model[bow_corpus]

### Running LDA using Bag-of-Words

Latent Dirichlet Allocation utilizes the tf-idf scores. 

The tf-idf scores rank the importance of a token in a review given a corpus of reviews.

The idea is that the more a token appears in a document, the higher the token is ranked (tf: term frequency).
However, the term gets a lower weight the more it appears in other documents (idf: inverse document frequency).

We played with the parameters to obtain a satisfying result (e.g., number of topics etc.).

# initiate model
lda_model = gensim.models.ldamodel.LdaModel(corpus=corpus_tfidf, 
                                            id2word=id2word,
                                            num_topics=6,
                                            random_state=42,
                                            passes=5,
                                            alpha='auto')

### Visualizing topics

# for visualization
pyLDAvis.enable_notebook()
vis = pyLDAvis.gensim.prepare(lda_model, bow_corpus, id2word, mds='mmds', R=15)

# visualize
vis

# check identified topics
for idx, topic in lda_model.print_topics(-1):
    print('Topic: {} \nWords: {}'.format(idx+1, topic))

### Running LDA using Bag-of-Words without tf-idf scores

Here, LDA is conducted with the bow_corpus and without tf-idf representation.

# initiate model
lda_model = gensim.models.ldamodel.LdaModel(corpus=bow_corpus, 
                                            id2word=id2word,
                                            num_topics=6,
                                            random_state=42,
                                            passes=5,
                                            alpha='auto')

### Visualizing topics

# for visualization
pyLDAvis.enable_notebook()
vis = pyLDAvis.gensim.prepare(lda_model, bow_corpus, id2word, mds='mmds', R=15)

# visualize
vis

# check identified topics
for idx, topic in lda_model.print_topics(-1):
    print('Topic: {} \nWords: {}'.format(idx+1, topic))
