==> Code/6_Naive_Bayes.ipynb <==
# Naive Bayes Classifier

### Import relevant libraries

# import relevant libraries

# for data manipulation
import numpy as np
import pandas as pd

# split data into training, validation and test set
from sklearn.model_selection import train_test_split

# for counter
from tqdm import tqdm

# for classifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.model_selection import GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix

# for oversampling
from imblearn.over_sampling import SMOTE

### Import data after pre-processing

# import the data
df = pd.read_csv('df_tokenized.csv')

# check shape
df.shape

### Split into training, validation and test set

##### Check dataframe

df.head()

df['sentiment'].value_counts()

y = np.where(df['sentiment']=='positive',1,0)

##### Split 60:20:20

# create test set
X_train, X_test, y_train, y_test = train_test_split(df['tokens'], y, test_size=0.2, stratify=y, random_state=42)

# create training and validation set
X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.25, random_state=42)

# check distributions
print(y_train.sum()/y_train.shape[0])
print(y_val.sum()/y_val.shape[0])
print(y_test.sum()/y_test.shape[0])

There is a strong imbalance, thus, some techniques against imbalance have to be applied.

# before that we will use grid-search 5-fold cross validation to find the optimum value for alpha therefore no validation set is needed
X_train = pd.concat([X_train,X_val])
y_train = np.append(y_train,y_val)

### Create feature space (DTM)

A computer cannot deal with raw text data, the texts have to be converted to feature space.

# initiate CountVectorizer
vectorizer = CountVectorizer()

# learn the vocabulary from training set
vectorizer.fit(X_train)

# create document-term matrix
X_train = vectorizer.transform(X_train)

# transform also test set to document-term matrix
X_test = vectorizer.transform(X_test)

# check the first features of the document-term matrix
vectorizer.get_feature_names_out()[1000:1010]

# check amount of unique vocabulary 
vectorizer.get_feature_names_out().shape

difference = y_train.shape[0]-y_train.sum()

y_train.shape

### Apply oversampling to tackle class imbalance 

##### Oversampling

Oversample the minority class so that there are more observations of it. Create as many so that half of the documents have negative sentiment.

# oversample the minority class
oversampler = SMOTE(sampling_strategy='minority', random_state=42)
X_train, y_train = oversampler.fit_resample(X_train, y_train)

# check new ratio of sentiment labels
print(y_train.sum()/y_train.shape[0])

### Build NB Classifier

First, build the Naive Bayes with Laplace smoothing.

# build Multinomial NB classifier, using a uniform prior distribution
nb_classifier = MultinomialNB(fit_prior=True)

# train the classifier
nb_classifier.fit(X_train, y_train)

# predict on the test set
y_pred = nb_classifier.predict(X_test)

# evaluate model performance
print("Accuracy: ", accuracy_score(y_test, y_pred))
print("Precision: ", precision_score(y_test, y_pred, average='macro'))
print("Recall: ", recall_score(y_test, y_pred, average='macro'))
print("F1: ", f1_score(y_test, y_pred, average='macro'))

# look at the classification report
report = classification_report(y_test, y_pred)
print("Classification Report:\n", report)

# look at the confusion matrix
confusion_matrix(y_test, y_pred)

# extract most important features for each label
for i, label in enumerate(nb_classifier.classes_):
    top = np.argsort(nb_classifier.feature_log_prob_[i])[-20:]
    print("Label %s: %s" % (label, ", ".join(vectorizer.get_feature_names_out()[top])))
    print()

##### Grid-search to optimize alpha value

# build Naive Bayes classifier
nb_classifier = MultinomialNB()

# define hyperparameters to tune
param_grid = {
    # alpha is the smoothing parameter for Naive Bayes
    'alpha': [1,5,10,15],
    'fit_prior': [True]
}

# grid search with 5-fold cross validation
grid_search = GridSearchCV(nb_classifier, param_grid, cv=5, scoring='f1_macro', refit=True, return_train_score=True, verbose=3)
grid_search.fit(X_train,y_train)

# check the models
pd.concat([pd.DataFrame(grid_search.cv_results_["params"]),pd.DataFrame(grid_search.cv_results_["mean_train_score"], columns=["Training Mean F1 Score"]),pd.DataFrame(grid_search.cv_results_["mean_test_score"], columns=["Validation Mean F1 Score"])], axis=1)

# print the best alpha value and f1 score
print("Best Hyperparameters: ", grid_search.best_params_)
print("Best F1 Score: ", grid_search.best_score_)

# predict on the test set
y_pred = grid_search.predict(X_test)

# look at the classification report
report = classification_report(y_test, y_pred, digits=4)
print("Classification Report:\n", report)

# look at the confusion matrix
confusion_matrix(y_test, y_pred)
