==> Code/9_GPT-3.5_evaluation.ipynb <==
Import packages

!pip install openai
import openai
import pandas as pd
import os
import re
from tenacity import retry, wait_random_exponential, stop_after_attempt
from tqdm import tqdm
import matplotlib.pyplot as plt
import numpy as np
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score

# GPT-3.5 evaluation

### Load data and get overview

# mount to Google Drive
from google.colab import drive
drive.mount('/content/drive')

# load the test set data
df = pd.read_csv('Handin/Data/df_test_set_sample.csv')

print(df.head())

print(df.shape)

# Directory to save the csv after each iteration, slow but secure
DIR_PATH = "/content/drive/MyDrive/NLP/Notebooks/Old"
CSV_FILE = os.path.join(DIR_PATH, 'gpt_responses.csv')

# Function to get the last saved index since the API is no stable
def get_last_saved_index():
    if os.path.isfile(CSV_FILE):
        df = pd.read_csv(CSV_FILE)
        return df.shape[0]
    return 0

# Function to process reviews
def process_with_retry():
    openai.api_key = "FILL IN AN VALID OPENAI KEY"

    # Get the index of the last saved file
    last_index = get_last_saved_index()
    if last_index > 0:
        df_gpt_response = pd.read_csv(CSV_FILE)
    else:
        df_gpt_response = pd.DataFrame(columns=['review_comment', 'predicted_rating', 'actual_rating'])

    # Iterate over each review comment in sample_df with a progress bar
    for i, (comment, actual_rating) in enumerate(tqdm(zip(df['reviews'][last_index:], df['sentiment'][last_index:]), desc="Processing reviews", unit="review")):
        response = openai.Completion.create(
            model="text-davinci-003",
            prompt=f"""Given the following Airbnb review, predict the reviewer's rating as 1 or 0,
            where 1 is positive and 0 is the negative. Provide your answer as only an integer. 
            Here is the review: '{comment}' """,
            temperature=0,
            max_tokens=60,
            logprobs=10,
            top_p=1.0,
            frequency_penalty=0.0,
            presence_penalty=0.0
        )

        try:
            # Extract the predicted rating from the response
            predicted_rating = int(response.choices[0].text.strip())

            # Append the review comment, predicted rating, and actual rating to the df_gpt_response DataFrame
            df_gpt_response = df_gpt_response.append({'review_comment': comment, 'predicted_rating': predicted_rating, 'actual_rating': actual_rating},
                                                     ignore_index=True)
            
            # Save DataFrame every iteration
            df_gpt_response.to_csv(CSV_FILE, index=False)
                
        except ValueError:
            # Handle cases where the predicted rating cannot be converted to an integer
            print(f"Error: Invalid response for comment - '{comment}'")
    
    return df_gpt_response

# Call the process_with_retry function and assign the result to a variable
result_df = process_with_retry()

# Access the df_gpt_response DataFrame
print(result_df)


### Get evaluation metrics



from sklearn.metrics import accuracy_score, classification_report

# Read the DataFrame
df_gpt_response = pd.read_csv("/Handin/Data/gpt_responses.csv")

# Extract the predicted and actual ratings
predicted_ratings = df_gpt_response['predicted_rating']
actual_ratings = df_gpt_response['actual_rating']

# Calculate the overall accuracy
accuracy = accuracy_score(actual_ratings, predicted_ratings)
print(f"Overall Accuracy: {accuracy}")

# Calculate precision, recall and F1-score for each category
classification_metrics = classification_report(actual_ratings, predicted_ratings, digits = 5)
print(classification_metrics)


### Get probabilities of the predicted tokens

openai.api_key = "FILL IN AN VALID OPENAI KEY"
# Make API request
FP = openai.Completion.create(
  model="text-davinci-003",
  prompt="""Given the following Airbnb review, predict the reviewer's rating as 1 or 0,
            where 1 is positive and 0 is the negative. Provide your answer as only an integer. 
            Here is the review: 'The flat was good, host was bad and location great' """,
  temperature=0,
  max_tokens=60,
  logprobs=10,
  top_p=1.0,
  frequency_penalty=0.0,
  presence_penalty=0.0
)

import math
# Remove spaces and all non numeric characters
def clean_number(x):
    x = re.sub("[^0-9]", "", x)
    return x

token_probas = FP["choices"][0]["logprobs"]["top_logprobs"]
probabilities = {str(k):0 for k in range(0,2)}

for token_prob in token_probas:
    for keys, probas in token_prob.items():
        num = clean_number(keys)
        if num in ["0","1"]:
            probabilities[num] += math.exp(probas)

# Normalize
total = sum(probabilities.values())
probabilities = {k: v / total for k, v in probabilities.items()}

print(probabilities)
